desktop: support infrared TV remotes
FIXME: this change is not tested yet!
       I can't test it from home, so
       it has to wait until I go into the office.

In PrisonPC images, it adds a layout for PrisonPC-branded remotes, and
hard-codes an assumption that EVERY remote is this layout (not strictly true).
This saves us having to work out what make/model all our existing remotes claim to be.

https://alloc.cyber.com.au/task/task.php?taskID=30362

lirc-era tickets (probably irrelevant):

https://alloc.cyber.com.au/task/task.php?taskID=30362
https://alloc.cyber.com.au/task/task.php?taskID=24557
https://alloc.cyber.com.au/task/task.php?taskID=24782