Add warning about resolved
Mike ran into this on his own system, where
(very late in the build!) it runs something like this

    # Now the initrd is built and copied out, remove its build dependencies.
    # Workaround stock kernels needing *a* linux-initramfs-tool.
    apt purge initramfs-tools tiny-initramfs+

He saw "deb.debian.org" fail to resolve.
This is because by that time,
inside the guest,
resolve.conf pointed to 127.0.0.53 due to this:

    --customize-hook=ln -nsf /lib/systemd/resolv.conf $1/etc/resolv.conf

Installing apt-cacher-ng worked as a workaround, because
now inside the guest apt finds and uses
http://127.0.0.1:3142 instead of http://deb.debian.org.

The other workarounds would be to

    1. make sure a DNS server is running on 127.0.0.53;
    2. switch resolve.conf to 127.0.0.53 later in the build; or
    3. avoid DNS resolution after resolve.conf moves.

Logically, #2 is the right thing to do, but
I don't really want to "break up" the
"install and configure basic networking" block of code, and
"enable resolved you chump!" (#1) is not too onerous.

Since it's not super obvious what has happened,
this warning is a reasonable reminder/memory jogger if this happens again.

This PROBABLY will only bite people building PrisonPC images, so
realistically it probably won't come up ever again.