inmate: set chromium proxy
Right now this only gets you as far as squid's password prompt.

This is because PrisonPC's squid requires a desktop login before it
will trust a browser login, and right now I haven't ported the desktop
login code over.

Previously this was using a .pac file bundled into the image.
The ONLY reason we EVER did that was because it's the only way to tell
chromium to talk https://proxy:3128 instead of http://proxy:3128.
That is, to encrypt the connection between the browser and the proxy.
This was a Good Thing because the user is sending a cleartext password
over it.

Unfortunately, SSL inspection (MITM attack / SSL bump) broke that.
So ever since we started SSL inspection, we've been doing cleartext
from the browser to the proxy.  This is pretty shit, but
(at least in theory) we can fall back on sites using port isolation
in their edge switches.  We also have a basic firewall on every host.
(The firewall isn't ported over yet, either.)