(BROKEN) pre-cache chrome components
This is not working.

If you drop unzipped .crx3 into /usr/share/chromium/extension/,
they load as extensions (not components).

If you drop unzipped .crx3 into /usr/share/chromium/components/,
I *THOUGHT* it was working, but I had accidentally unblocked direct
internet access, so chromium was just downloading direct from google.

I *THINK* we have to have an autostart job that copies content into
~/.config/chromium/Default/ on login, but I don't know how do deal
with the manifest files.  I suspect that metadata has to go into
somewhere special.

This functionality is important for security, BUT
skipping it is NO WORSE than the Debian 9 SOEs, so
I'm leaving this for now and hopefully I can come back to this later.