Move web apps out of the Start menu and into chromium Apps menu
Not moving Mail yet because chromium shows that as "insecure" because there is no https,
even though it is running on localhost.  Because it runs on the desktop, to do https,
we would have to use the snakeoil certs created by ca-certificates package.
Which are trusted automatically by gnutls and openssl, but NOT by FUCKING CHROMIUM AND ITS FUCKING NSS3.
To make chromium trust it, we would have to ship certutil(1nss-tools) and have it patch ~p123/.pki/ on EVERY BOOT.
We used to do exactly that, and we know that it leads to permanent corruption in ~/.config/chromium unless the user is prevented from running chromium until it finishes.
It is just... really awful.
We also do not really want to ship certutil in the first place since it is a low-level cert management tool.