delete-bad-files: find .glob relative to argv[0] (not hard-coded)
The purpose of this is so I can symlink the script into another dir, and
have it read the .glob next to the symlink.

This in turn will let me have separate "only inmates" and "everyone" globs.

This in turn will let me ship ssh/sftp in staff images.

This in turn will let me use sftp:// in gvfs (which just fork+exec's /bin/ssh, because sigh.