Bugfix: SHOULD/MUST and SHOULD/MUST NOT overlap: {'RANDOMIZE_BASE'}
RANDOMIZE_BASE causes KASLR i.e. makes the offsets of specific parts of the kernel difficult to guess in advance.
This is a cheap-and-powerful hardening feature.  We want it.  I do not know why we had it in both SHOULD and SHOULD NOT before.
I think this was the case for a long time and we did not notice because until today, there was no reporting for overlaps.