kernel: create metapackage also
This means where the old bootstrap.git did hard-coded:

    ln -s initrd.img-5.4.13inmate $t/boot/initrd.img
    ln -s    vmlinuz-5.4.13inmate $t/boot/vmlinuz
    apt install linux-image-5.4.13inmate

Now we can just do

    apt install linux-image-inmate

And it will pick the latest-built one.

Note that this means if you build an OLD version of this repo,
it will still use the latest kernel in the prisonpc apt repo!

To cherry-pick an old version (like the bootstrap.git behaviour),

    apt install linux-image-inmate=5.4.3inmate